// src/app/shared/models/artical.ts
export interface IArtical{
    id: 0;
    address: string;
    description: string;
    img_path:string;
}

// "id": 2,
//         "img": "Article-60699f102deef.jpg",
//         "address": "dfydjhgsd sdhs dh fd4f45d77474",
//         "description": "ddhdhdggdueuv iehwiw wehiweh ihwei wihfiwh fiwehfiehwi fi fee fifhhf hfihf ihwifh wfiw ifhhfe h whfwef fhehf ewhfeh",

//         "img_path": "http://127.0.0.1:8000/uploads/article/Article-60699f102deef.jpg"